requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
numpy==1.26.4
pillow==10.3.0
tflite-runtime==2.14.0
